refactor lp10-parentI2 so that it can be used by lp16
