+++ tests/unit/libpluto/lp10-parentI2/output1.txt
@@ -1,6 +1,8 @@
 ./parentI2 ike_alg_register_enc(): Activating OAKLEY_AES_CBC: Ok (ret=0)
 ./parentI2 ike_alg_register_hash(): Activating OAKLEY_SHA2_512: Ok (ret=0)
 ./parentI2 ike_alg_register_hash(): Activating OAKLEY_SHA2_256: Ok (ret=0)
+./parentI2 loading secrets from "../samples/parker.secrets"
+./parentI2 loaded private key for keyid: PPK_RSA:AQN7wUerV
 | processing whack message of size: A
 | processing whack message of size: A
 processing whack msg time: X size: Y
@@ -602,6 +604,8 @@ sending 556 bytes for STATE_PARENT_I1 through eth0:500 to 132.213.238.7:500 (usi
 ./parentI2 leak: default conf conffile, item size: X
 ./parentI2 leak: default conf ipsecd_dir, item size: X
 ./parentI2 leak: default conf ipsec_conf_dir, item size: X
+./parentI2 leak: 2 * id list, item size: X
+./parentI2 leak: secret, item size: X
 ./parentI2 leak: 2 * hasher name, item size: X
 ./parentI2 leak detective found Z leaks, total size X
 Pre-amble (offset: 65): #!-pluto-whack-file- recorded on FOO
