+++ tests/unit/libpluto/lp10-parentI2/parentI2.c
@@ -1,140 +1,23 @@
-#define LEAK_DETECTIVE
-#define AGGRESSIVE 1
-#define XAUTH
-#define MODECFG
-#define DEBUG 1
-#define PRINT_SA_DEBUG 1
-#define USE_KEYRR 1
-
-#include "constants.h"
-#include "oswalloc.h"
-#include "whack.h"
-#include "../../programs/pluto/rcv_whack.h"
-
-#include "../../programs/pluto/connections.c"
-
-#include "whackmsgtestlib.c"
-#include "seam_timer.c"
-#include "seam_fakevendor.c"
-#include "seam_pending.c"
-#include "seam_ikev1.c"
-#include "seam_crypt.c"
-#include "seam_kernel.c"
-#include "seam_rnd.c"
-#include "seam_log.c"
-#include "seam_xauth.c"
-#include "seam_host_parker.c"
-#include "seam_terminate.c"
-#include "seam_x509.c"
-#include "seam_spdbstruct.c"
-#include "seam_demux.c"
-#include "seam_commhandle.c"
-#include "seam_whack.c"
+#include "parentI2_head.c"
 #include "seam_keys.c"
-#include "seam_exitlog.c"
-#include "seam_natt.c"
-#include "seam_dnskey.c"
+#include "seam_host_parker.c"
 
-u_int8_t reply_buffer[MAX_OUTPUT_UDP_SIZE];
+#define TESTNAME "parentI2"
 
-#include "seam_gi_sha1.c"
-#include "seam_gi_sha1_group14.c"
-#include "seam_finish.c"
-
-#include "ikev2sendI1.c"
-
-/* this is replicated in the unit test cases since the patching up of the crypto values is case specific */
-void recv_pcap_packet(u_char *user
-		      , const struct pcap_pkthdr *h
-		      , const u_char *bytes)
+static void init_local_interface(void)
 {
-    struct state *st;
-    struct pcr_kenonce *kn = &crypto_req->pcr_d.kn;
-
-    recv_pcap_packet_gen(user, h, bytes);
-
-    /* find st involved */
-    st = state_with_serialno(1);
-    if(st != NULL) {
-        passert(st != NULL);
-        st->st_connection->extra_debugging = DBG_EMITTING|DBG_CONTROL|DBG_CONTROLMORE|DBG_CRYPT|DBG_PRIVATE;
-    }
-
-    run_continuation(crypto_req);
-}
-
-main(int argc, char *argv[])
-{
-    int   len;
-    char *infile;
-    char *conn_name;
-    int  lineno=0;
-    int  regression = 0;
-    struct connection *c1;
-    struct state *st;
-
-    EF_PROTECT_FREE=1;
-
-    progname = argv[0];
-    leak_detective = 1;
-
-    /* skip argv0 */
-    argc--; argv++;
-
-    if(strcmp(argv[0], "-r")==0) {
-        regression = 1;
-        argc--; argv++;
-    }
-    if(argc != 4) {
-	fprintf(stderr, "Usage: %s [-r] <whackrecord> <conn-name> <pcapfile> <pcapout>\n", progname);
-	exit(10);
-    }
-
-    tool_init_log();
-    init_crypto();
-    load_oswcrypto();
-    init_fake_vendorid();
     init_parker_interface();
-
-    infile = argv[0];
-    conn_name = argv[1];
-
-    cur_debugging = DBG_CONTROL|DBG_CONTROLMORE;
-    if(readwhackmsg(infile) == 0) exit(10);
-
-    /* input packets */
-    recv_pcap_setup(argv[2]);
-
-    /* output first packets to /dev/null */
-    send_packet_setup_pcap("/dev/null");
-
-    c1 = con_by_name(conn_name, TRUE);
-    assert(c1 != NULL);
-
-    assert(orient(c1, 500));
-    show_one_connection(c1);
-
-    st = sendI1(c1, DBG_CONTROL, regression == 0);
-
-    /* now accept the reply packet:
-       output interesting packet to capture file
-    */
-    send_packet_setup_pcap(argv[3]);
-
-    cur_debugging = DBG_CONTROL|DBG_CONTROLMORE|DBG_PARSING;
-    pcap_dispatch(pt, 1, recv_pcap_packet, NULL);
-
-    st = state_with_serialno(1);
-    if(st!=NULL) {
-        free_state(st);
-    }
-
-    report_leaks();
-
-    tool_close_log();
-    exit(0);
 }
 
+static void init_fake_secrets(void)
+{
+    osw_load_preshared_secrets(&pluto_secrets
+			       , TRUE
+			       , "../samples/parker.secrets"
+			       , NULL);
+}
+
+#include "parentI2_main.c"
 
  /*
  * Local Variables:
