+++ tests/unit/libpluto/lp10-parentI2/parentI2_head.c
@@ -27,6 +27,7 @@
 #include "seam_x509.c"
 #include "seam_spdbstruct.c"
 #include "seam_demux.c"
+#include "seam_commhandle.c"
 #include "seam_whack.c"
 #include "seam_exitlog.c"
 #include "seam_natt.c"
@@ -40,10 +41,3 @@ u_int8_t reply_buffer[MAX_OUTPUT_UDP_SIZE];
 
 #include "ikev2sendI1.c"
 
- /*
- * Local Variables:
- * c-style: pluto
- * c-basic-offset: 4
- * compile-command: "make check"
- * End:
- */
